Remove extra file
