+++ api4/system_local.go
@@ -3,7 +3,35 @@
 
 package api4
 
+import (
+	"encoding/json"
+	"net/http"
+
+	"github.com/mattermost/mattermost-server/v5/audit"
+	"github.com/mattermost/mattermost-server/v5/model"
+)
+
 func (api *API) InitSystemLocal() {
 	api.BaseRoutes.ApiRoot.Handle("/logs", api.ApiLocal(getLogs)).Methods("GET")
 	api.BaseRoutes.ApiRoot.Handle("/integrity", api.ApiLocal(localCheckIntegrity)).Methods("POST")
+}
+
+func localCheckIntegrity(c *Context, w http.ResponseWriter, r *http.Request) {
+	auditRec := c.MakeAuditRecord("localCheckIntegrity", audit.Fail)
+	defer c.LogAuditRec(auditRec)
+
+	var results []model.IntegrityCheckResult
+	resultsChan := c.App.Srv().Store.CheckIntegrity()
+	for result := range resultsChan {
+		results = append(results, result)
+	}
+
+	data, err := json.Marshal(results)
+	if err != nil {
+		c.Err = model.NewAppError("Api4.localCheckIntegrity", "api.marshal_error", nil, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	auditRec.Success()
+	w.Write(data)
 }
